requirements.txt
discord.py
openai
customtkinter
psutil
pynvml
httpx